Test: Fix map_t_cose_errors()
C arrays start from index 0, so the correct test for "does this entry
exist?" is "< size" (or ">= size").
Also the mapping array t_cose_verify_error_map was missing values for a
number of enumerals from enum attest_token_err_t, one of which was in
the middle, meaning that the remaining entries were all in the wrong
places.

Replace the difficult-to-maintain array with a switch statement.

Change-Id: I24a62c732848c23e6b30a8e21bb9c05e8c5e023f
Signed-off-by: Chris Brand <chris.brand@cypress.com>