Platform: Set OVERWRITE_ONLY as default on S1/B1
Remove the upgrade strategy override from the Musca-B1 and S1
CMake files, effectively setting OVERWRITE_ONLY as the default.

Increase the size of the trailer so the increasingly large TLV
area won't cause the signed image to exceed the partition size.
Update flash layout in order to support the SWAP upgrade strategy.

Change-Id: I1cd097d153d9de4dad5dceeb4ef5bca21d67bc26
Signed-off-by: Balint Matyi <Balint.Matyi@arm.com>