PSoC64: Support runtime specification of ITS/PS flash
Add a set of flags that can be used to indicate "this SMPU region
should be configured based on provisioning data".
Use those flags for both the ITS and PS SMPU configurations.
Add SMPU configuration code to detect those flags and, if set, to
call tfm_hal_its_fs_info() or tfm_hal_ps_fs_info() to retrieve the
base address and size of the ITS flash region, and to configure the
SMPU accordingly.
Various follow-on changes to the code to dump the SMPU config.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I2ad9b4ff5f8235c8141ae337a85815d71669fe8f