QCBOR: Avoid apparent null pointer dereference
In GetNext_FullItem(), it appears possible to get to the Done label
with pAllocator set to NULL (for example, it could be set to NULL
when first declared, and then the first "goto Done" could be
reached. In this case, StringAllocator_Free() will attempt to
dereference the NULL pointer.

Move the Done: label to avoid this.

Change-Id: I07502336116d39d4d030e9cee656481d92d5bda6
Signed-off-by: Chris Brand <chris.brand@cypress.com>