Core: Fix 'event' malfunction behaviour
The 'event' object is designed as can be waited by one owner thread,
so the 'event' object re-use owner thread's state as its own state to
save object size. Since one thread can be owners of multiple events,
this re-using can lead to a malfunction that another unexpected event
would wake up instead of the expecting one, since the unexpected event
has the same owner with the expecting event.

Clear the owner while an event wakes up, this could ensure the state
of events is not connected by the same owner's state.

Change-Id: I8e13d211de8b952cc429e3dadd0e73f6205b47a5
Signed-off-by: Ken Liu <ken.liu@arm.com>