Build: Add support for specifying GNUARM_PREFIX
Add support to allow for vendor specific GNU triplets that may not be
arm-none-eabi, but something like arm-<VENDOR>-eabi.

In cmake/Compiler/GNUARM.cmake we provide a means to extract the
compiler prefix from CMAKE_C_COMPILER as it will not always be set.  For
example when mbed-crypto's CMake is invoked from tfm, we'll eventually
source GNUARM.cmake and not have GNUARM_PREFIX set.  This provides a
means to handle such cases (otherwise we end up with NOTFOUND cache
entries for CMAKE_LINKER & CMAKE_AR.

Change-Id: I5b3f1a0ff1156b9eda6c540220aa99162b6fd813
Signed-off-by: Kumar Gala <kumar.gala@linaro.org>