Tools: Use build system appropriate newlines
After running tools/tfm_parse_manifest_list.py on Windows, git lists all
generated files as changed, even if it is just the newlines that has
changed.

git converts the newlines on checkout and commit, so the files should
really have newlines that are appropriate for the build system.

This change makes python generate the appropriate newlines so that only
truly changed files will be listed by git.

Signed-off-by: TTornblom <thomas.tornblom@iar.com>
Change-Id: Iba0a0a7f162de432f2c5eba938539759e00dd376