Build: Replace CMSIS_5_DIR with deeper CMSIS_DIR
This patch replaces the CMake variable CMSIS_5_DIR with the deeper
CMSIS_5_DIR/CMSIS.
This is the preparation for the RTX references from CMSIS_5 repo
to tf-m-tests repo. Because setting 'CMSIS_5_DIR' to 'tf-m-tests'
could be confusing.
With this change, the 'CMSIS_DIR' would be set to 'tf-m-tests/CMSIS'
later.

Change-Id: I3bf19d15592a24cb718b220d8efc49f42b40f56d
Signed-off-by: Kevin Peng <kevin.peng@arm.com>