Build: Use CMSIS for testing from tf-m-tests
TF-M testing is using CMSIS RTX as NS OS. It links to the
pre-built RTX libraries from the CMSIS_5 repo.
Because the CMSIS_5 repo will remove the libraries, a copy
of necessary RTX libraries along with source codes is stored
in the tf-m-tests repo.
This patch is to move the CMSIS RTX reference from CMSIS_5
to tf-m-tests.

Change-Id: I19e9ac333b200e69a749607fd1a2374805861be8
Signed-off-by: Kevin Peng <kevin.peng@arm.com>