Platform: PSoC64: Power failure NV counter protection
Add code to protect the (flash) NV counters from power
failures. The goal is to ensure that after a write,
even if the power fails at some point during the write,
we will re-initialise such that a read will always get
either the old (pre-write) value or the new (post-write)
value, but never anything else.

To do this, we allocate two flash sectors for
NV counters, with the second acting as a backup. When
writing, we first erase and then write the second (backup)
sector before erasing the first (primary) sector.

In addition, to help protect against power failures during
a write or erase, when the state is completely unpredictable,
we also add a checksum for the sector.

A sector is assumed to contain valid counters if both the
"watermark" is present and the checksum matches.

During initialisation (i.e. power restoration), if the primary
sector is not valid, we check the backup sector. If the backup
sector is valid, we erase the primary sector and copy the
backup sector to it. If the backup sector is also not valid,
we continue with the existing initialisation code.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: Ibeabdfc23b138fe7ef8f4e604b7906ce2e829e42