Platform: Disallow repeated provisioning on Musca-S1
The Musca-S1 does not have a real OTP, so its crypto hardware values
are saved to MRAM instead.
Before, the life-cycle state was only checked on the S1 from the MRAM
once provisioning was already done. This resulted in being able to run
provisioning repeatedly. Change this so that the values in the MRAM are
used always, so that one needs to wipe the MRAM to repeat the process.
Also, update the relevant docs.

Signed-off-by: Balint Matyi <Balint.Matyi@arm.com>
Change-Id: I51d36b6b453ccfd1770955a4fe3f5f6fdfd70cf0