Test: Warn when a write-once UID already exists
Changes the storage tests to print a warning, but not fail the test,
when the UID that is created with the PSA_STORAGE_FLAG_WRITE_ONCE flag
already exists in storage.

This makes it possible to run the storage tests multiple times with
persistent storage without having failing tests. This is more desirable
than automatically wiping the storage before every test run, as it
provides proper test coverage for initialisation with an existing
filesystem.

Change-Id: Ia25384b6a1bb7fc0c718752b9a1ce409f1969507
Signed-off-by: Jamie Fox <jamie.fox@arm.com>