Test: Remove warning when testing persistent storage
Removes the warning that is displayed when the storage tests are run
with persistent storage, rather than the RAM FS. Testing with
persistent storage can be desirable, and wearing out the flash may not
be a concern for a given device. The RAM FS configuration is still
displayed at build time, so the test-runner may make their own
conclusions.

Change-Id: I6b9d5ce354cc6ddfd681e9387cdb852a5705f861
Signed-off-by: Jamie Fox <jamie.fox@arm.com>