Build: Disable RAM FS by default
Changes the common configuration to disable the ITS and PS RAM FS by
default, even when running the regression tests. Testing with persistent
storage allows for greater coverage of scenarios such as initialisation
with existing storage and correct erasing of flash before overwriting.

The setting can still be overridden in the platform CMake config, or by
providing it on the command line if flash wear is a concern.

Change-Id: If950e13d5f89ff164b47ed8d6602cdef5d4d6b5e
Signed-off-by: Jamie Fox <jamie.fox@arm.com>