ITS: Combine filesystem create and write functions
Merges the filesystem create function into the write function. Adds
flags to indicate that a file should be created if it does not already
exist and that any existing file should be truncated and replaced.

This makes it possible to do atomic replacement of existing files,
while also reducing code duplication in the filesystem

Change-Id: Ie6fcc8081b854738aae45a4e0f754ab80fb0178b
Signed-off-by: Jamie Fox <jamie.fox@arm.com>