ITS: Pad ITS structs to ITS_FLASH_MAX_ALIGNMENT size
This is a refactoring of some ITS related structs that needs to
be sized appropriately for flash writing.

Originally the types used alignment to have the proper size, but
the IAR compiler can not have have automatic variables with stricter
alignment than the stack, which has an alignment of 8. This patch
instead adds padding at the end of the structs if necessary, which should
save a few bytes of stack space as well as work on all toolchains.

Signed-off-by: TTornblom <thomas.tornblom@iar.com>
Change-Id: I76a2dfabd37fe1fdb4338218f5da2c50176032c4