Attest: Remove LEGACY_TFM_TLV_HEADER support
The boot record is shared between bootloader and
runtime firmware in the structure of TLV data.
The header of the TLV fields are filled differently
by original MCUboot and its TF-M fork. The config
option LEGACY_TFM_TLV_HEADER was meant to make the
distinction. Since the forked version of MCUboot is
going to be removed therefore this differentiation
could also be deleted.

Signed-off-by: Balint Matyi <Balint.Matyi@arm.com>
Change-Id: Ie3904996eb0f1006eaa67c324dfbbb21a9e146ab