Test: Remove test from TF-M
This patch removes the test code in TF-M. The codes should have
been moved to the tf-m-tests repo.

This patch also contains changes to make compilation work:
1. Update the TEST_DIR
2. Update manifest list and re-generate the manifests
3. Add necessary include path to build

Note:
This patch only intends to remove the test codes.
There are some intentional "workaround like" changes such as the
include paths.
They will be refined in the future.

Change-Id: I8d301e1d8f4fc7818b45c1b5ebf938a6b721f061
Signed-off-by: Kevin Peng <kevin.peng@arm.com>