Platform: Simplify ARMCLANG - IAR compatibility
ARMCLANG defines a number of special symbols, not normally defined
by IAR (or GCC). The IAR linker has no facility to define these
symbols in the linker script, like the Gnu linker. These symbols used
to be defined on the command line to the compiler through CMake, but it
was causing issues where bash was needed to execute the commands and it
was impossible to define these symbols in the IDE. Defining them in
region.h instead solves this.

Signed-off-by: TTornblom <thomas.tornblom@iar.com>
Change-Id: Iaffe21ac530df3316b1bf71aad392262f9cf8017