SPM: Refine tfm_spm_init code
This patch refines the tfm_spm_init code.
The intention is to make it easy to apply the HAL API later.
With the new HAL API, the SPM would not have the partition database.
And should use partition structures for initialisation.

This patch also refines the tfm_spm_partition_get_running_partition()
and tfm_spm_partition_get_running_partition_id() implementations:
The running parition is actually got first, then the ID.
So the codes are "swapped" between the two functions.

Change-Id: I8407c77706d27a17b9347258cb6299ea34e962b6
Signed-off-by: Kevin Peng <kevin.peng@arm.com>