Docs: Rename folders for IPC and library model
As most of the files are named as xxx_func or xxx_ipc, rename the folder
to follow this default style, which makes the looking more aligned.
The library model also being mentioned as the function model, and the
IPC model is a PSA FF complied model. This explains why the suffix is
'psa' and 'func'.

Signed-off-by: Ken Liu <ken.liu@arm.com>
Change-Id: Id47e90624550ba038ed7b30bf494d88acde1b7be