SPM: Fix the sfn parameters check function of Lib Model
The tfm_core_check_sfn_parameters() function assumes all the secure
service callers to be unprivileged mode, which is not correct.

The privileged mode of NS Secure Service caller will be decided in
tfm_core_has_xx_access_to_region() functions, that's fine.

But the Secure caller can be only privileged mode because the whole
SPE is running under privileged mode.

This patch addresses this issue.

Change-Id: I16e016f1837d29bd0009f4404e02cb3b51732c45
Signed-off-by: Kevin Peng <kevin.peng@arm.com>