Platform: Make possible to set flash device for NV counters
If a platform uses the template NV counters implementation and have
more than one Flash device it can happen that firmware images and
NV counters are allocated to different flash areas. So the flash
device for NV counters can be set by the NV_COUNTERS_FLASH_DEV_NAME
macro. If it is not set, the FLASH_DEV_NAME macro will be used as the
flash device.

Change-Id: I2caf2984218994d8e79ed3f0cfba5a959c3f76b4
Signed-off-by: Mark Horvath <mark.horvath@arm.com>