Stack_seal_mitigation: Seal the PSP stacks for Library mode.
This patch does the changes for sealing the PSP stacks used
in library mode. There are 2 stacks used as PSP in Library mode :
ARM_LIB_STACK and TFM_SECURE_STACK. In case of ARM_LIB_STACK,
a new section is created for sealing the stack. The TFM_SECURE_STACK
needs more careful manipulation as the top of stack is used for iovec
parameters. So the code inserts a stack seal between the iovec
parameters and Partition stack area.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I028c625e53a8bb1677a7f3001095ab3452f82c32