Crypto: Correct setting of config parameters
This patch corrects the previously incorrect setting
of config parameters via cmake. The patch also displays
the values of the config parameters during build to
ease analysis of the config values being applied.

The crypto module now assumes that the config parameters
are passed in via build and hence the previous assumption
of setting the defaults in code are removed and #error
is introduced if the build doesn't set them.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Ic4dc8b2a08be5aa278d54cf4f4b73c0e12e9ccd3