Docs: Updating references to DS-5, and other content
This patch is updating the references to DS-5 to
Arm Development Studio. The version numbers,
and user instructions have also been updated
to reflect the new name.

It is also adding proper documentation to the
appropriate arm developer pages for each toolchain,
as well as fixing other minor formatting issues
across documentation

Signed-off-by: Minos Galanakis <minos.galanakis@arm.com>
Change-Id: I99fd4ff3f127e6fe9f1f5384245770637fe41c11