Platform: Adding Nuvoton M2351 Support
The NuMicro® M2351 microcontroller series is powered by Arm® Cortex®-M23
core with TrustZone® for Armv8-M architecture, which elevates the
traditional firmware security to the new level of robust software
security.

The porting will apply trusted-firmware-M to M2351 to make it easy to
create the security firmware on Nuvoton M2351 platform.

Signed-off-by: WSChang <wschang0@nuvoton.com>

Change-Id: I059618223fae8ab9ca085e9c998d51e3d12e8d86