Platform: Enable multiple data and bss section support
Enable the multiple data section copy and multiple bss
section clear on platform level. It always enabled in
case of secure firmware because there are data and bss
section per service.

In case of bootloader, the following platforms need this feature.

On Musca-B1 two section must be copied to SRAM by startup code:
- data section
- eFlash driver, it is executed from SRAM.

On Musca-A the entire bootloader copied and executed from SRAM
plus the data section also copied.

Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Change-Id: Ibf91a656d5897b66bfbb57e6d6c61b08347b3a65