Platform: Adjust the location of MT25QL_DEV_NS on Musca-A
The variable MT25QL_DEV_NS is placed in TFM_SP_PS_ATTR_RW section
which is not explicitly described in musca_bl2.ld. As a result,
the variable is placed after the data section. But the data
section in the copy table does not cover the data after the data
section. This issue causes the HardFault in BL2.

As MT25QL_DEV_NS is not part of the PS partition. So, move the
variable into data section directly.

Change-Id: I335bae2d60903c4e4d9afcf9d1dab8860ea2722c
Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>