Platform: Add support for nRF5340 DK board
The nRF5340 DK (Development Kit) board differs from the already
supported nRF5340 PDK (Preview Development Kit) with the revision
of the nRF5340 SoC. The revision used in PDK, called Engineering A,
has bigger size (32 kB instead of 16 kB) of IDAU flash regions and
several peripherals, e.g. USBD, are not functional on it.
This commit introduces definition of the nrf5340dk_nrf5340_cpuapp
platform that is almost the same as the nrf5340pdk_nrf5340_cpuapp
one, and it adds the NRF5340_ENG_A macro to the definition of the
latter for handling differences between those two in the shared code
contained in the common/nrf5340 directory.
Additionally, the missing entry for the NRF_FPU peripheral is added
in the initial peripheral configuration of the nrf5340 target.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
Change-Id: Ibc06690c216c1792ff03f74c284aeb6ac8326d53