Platform: NXP: Fix common USART driver
The GetTxCount and GetRxCount functions in the common USART Driver
should have returned an unsigned value according to the CMSIS-Driver
specification, but instead they have returned the negative
ARM_DRIVER_ERROR_UNSUPPORTED error code which can't be handled in a
standardized way and remained hidden.

This issue resulted in the SPM log messages appearing incorrectly.

Change-Id: Iaa89154520c4f87b4baa0672628c5b0036808a23
Signed-off-by: David Vincze <david.vincze@linaro.org>