Plaform: Fix NRF5340_ENG_A definition in the nRF5340 PDK target
This is a follow-up to commit bb4d5c59c02c1df0cc3baa393d18645c090766e2.

The nRF5340 PDK target requires the macro NRF5340_ENG_A to be defined
so that its nRF5340 SoC revision (Engineering A) is handled correctly
in the common nRF5340 code. This macro needs to be defined also during
linking, as it is used in the region_defs.h file included from the
linker script. Thus, use the `add_definitions` statement to define the
macro globally, not only for compilation of the secure and non-secure
images.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
Change-Id: I089fcdfafd2eaccab3baebbbaecdd824f72d0881