trusted-firmware-m: CMakeLists update after bumping to TF-Mv1.2.0-RC2
Required updates in CMakeLists.txt due to minor modifications in
the build CMAKE variables, as a result of bumping TF-M to
the TF-Mv1.2.0 RC2 tag.

The usage of CROSS_COMPILE allows us to specify the path to the
toolchain which Zephyr uses, so we do not need any downstream-
only patches for this. Also, the GNUARM variable is not needed
anymore and it is removed.

The commit also removes the cmake_minimum_required, since this
external project is supposed to be used with Zephyr samples,
tests, or applications, which shall define the minimum required
cmake version.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>