Platform: Place global variables in shared code to dedicated section
The code sharing between bootloader and runtime firmware requires to
share the global variables as well. Because the RAM area is reused by
SPE when bootloader has finished its job, therefore it must be ensured
that global variables are placed by the linker to the same place during
both linking operation (MCUboot and SPE). This location is a dedicated
section which has a fixed place at the beginning of the RAM and hence
not colliding with other sections. If code sharing is disabled then this
special section is not created.

Change-Id: I4b6c181f924c1ab8c1733b99806e824526fb820b
Signed-off-by: Tamas Ban <tamas.ban@arm.com>