Build: Fetch tf-m-tests repo with fixed version
The build system currently fetches the latest tf-m-tests repo always.
The build would be broken if there was a dependency change between
TF-M repo and test repo and developer was working on a older TF-M
commit.

This patch changes the version of tf-m-tests to be fetched to specific
commits.

With this patch, the build system will download a fixed workable version of
tf-m-tests.

Documentation on how the version will be updated is also added.

Change-Id: Ia6fbf824f1cc3e3509aebe3ad71b04291a81c8a2
Signed-off-by: Kevin Peng <kevin.peng@arm.com>