PSoC64: Fix IRQ test (TFM_CORE_TEST_1017)
---- Problem ----
IRQ test has problem to run on Isolation Level 2

---- Info ----
TFM_CORE_TEST_1017 is part of the Core non-secure tests.
Set ON to the following two MACROs to enable the test.
  TEST_NS=ON
  TFM_IRQ_TEST=ON

---- History ----
The initial implementation was at the time when the
isolation level was level 1.  There was no problem in the
original design to share configuration data between
TIMER0 (Secure) & TIMER1 (Non-Secure).

The data sharing between NS & S triggers the problem when
the development moved to Isolation Level 2.

---- Solution ----
Separate the data for TIMER0 & TIMER1

Change-Id: I1f0a0c4ae09356c3dff8378202e01dccf866473d
Signed-off-by: Alamy Liu <alamy.liu@infineon.com>