Boot: Move Mbed TLS configuration file for MCUboot
Move the configuration template to a standalone 'config' directory.

Enables support for CMSIS-Packs.

Avoids that the configuration template from the pack is in the include search
path together with other include files from the same 'include' directory.

Configuration file is copied to the project and its directory is added
to the include search path.

Signed-off-by: Robert Rostohar <Robert.Rostohar@arm.com>
Change-Id: I39afeafdcbf26e83a73c711cf1ae86e5e2667b1d