Build: Fix MCUboot image dependency checking
Previously the secure image was depeding on the version of the secure
image, and the NS on the NS. Now the NS correctly depends on S and S
depends on NS. Update the documentation strings of the image dependency
cmake options to be more clear. Remove dependency in image when
MCUBOOT_IMAGE_NUMBER == 1 as it isn't checked in that case.

Change-Id: I32557678cc9cdc8485fb812b2b08e5667b2fd9c0
Signed-off-by: Raef Coles <raef.coles@arm.com>