SPM: Refine list operations
List operations is necessary for FFM common logic. Move the header
into 'ffm' folder, and covert inline funtions into MACROs, as list
operations are fast and simple enough.

The main usage now for the lists is a ring based on the bidirectional
list structure. Other list usage can be added into this file when
necessary.

Change-Id: I45c5eab864f5c4e261b39c89cd42325bb98fa17c
Signed-off-by: Ken Liu <Ken.Liu@arm.com>