Platform diphda: Add OpenAMP
The ARM's Diphda platform uses openamp for tf-m non-secure communication.
The openamp interface is used to recieve messages and send response to
the host (Linux). The openamp library, which is part of platform code,
sits between openamp and SPM and uses TF-M RPC interface for SPM
interaction.

TF-M has Mailbox for non-secure communication but it is not supported
for Linux based non-secure side. In the Diphda, the non-secure side is
Linux environment so openamp is used instead.

The file naming convention used here is aligned with TF-M's secure_fw
naming convention.

The OpenAMP and Libmetal are added as cmake external projects and this
patch uses standard APIs of both the projects.

Change-Id: Id7f4d031b97ab8bc9005dfc7365c37b9e062ec9b
Signed-off-by: Mark Horvath <mark.horvath@arm.com>
Signed-off-by: Satish Kumar <satish.kumar01@arm.com>