platform: Add deprecation warnings for nRF5340 PDK
Indicate that the nRF5340 PDK (nordic_nrf/nrf5340pdk_nrf5340_cpuapp)
platform is deprecated by issuing a corresponding cmake warning at
build time and a run-time log message when the platform is initialized.

Add also the platform to the List of Deprecated Platforms in the
Platform deprecation and removal section of the TF-M documentation.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
Change-Id: Ibe657cf03d8973703418ac260a8d20d7fa1a85d7