SPM: Check both MODE and SPSEL to choose the stack during SVC
The exception frame is never placed on PSP when in handler mode.
If SPSEL is 1 (PSP), it is "read as 0", which means that the SVC handler
must also check the MODE bit of the EXC_RETURN to properly know which
stack the exception frame is on.

Previously, the checking of the SPSEL bit was done in assembly in a number
of places. This patch refactors that logic out of assembly, into C code,
since the new logic is more complicated.

Note that the logic works just as well for ARMv7 and below.

This patch is initially created by:
Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>

Change-Id: I62dda123d7a410251091c416954d2ac4904568ba
Signed-off-by: Kevin Peng <kevin.peng@arm.com>