SPM: Update thread state naming
The thread state THRD_STATE_RUNNING is actually used as runnable.
For example, in the tfm_event_wake() function, a thread woke up by
an event is only becoming runnable. Whether to make it as the current
running thread is decided by scheduler with other condition checks
such as priority - only the thread has highest priority in the runnable
list can run.
The current running thread is indicated by CURR_THRD.

This patch updates the naming to reflect the actual implementation.
Also moves some static functions to caller to simply the code.

Change-Id: Ib8c843c85feff35d6d6287bf82093801e3168e80
Signed-off-by: Kevin Peng <kevin.peng@arm.com>