SPM: Wake up only when waitfing for the signal
If a Partition is not waiting for a signal, it should not be woken
up. Appending the signal to its asserted_signals is enough. The next
time it waits for this signal, it will get it.
And the waiting signal should be cleared when waking up.
Setting waiting signal for polling mode is not necessary either.

Change-Id: I9929fd7a4081077eb9543ef3b95ac7278931496c
Signed-off-by: Kevin Peng <kevin.peng@arm.com>