SPM: Convert SVC number to uint8_t
SVC number encoded in SVC instruction is 8-bit long.
Currently it relies on the short-enum compiler option to have a
8-bit long SVC number type.

This patch converts the enum to uint8_t for SVC number and divids
the SVC numbers to two parts for IPC model:
 - 0x0 ~ 0x7F for SVC calls only allowed from Thread Mode
 - 0x80 ~ 0xFF for SVC calls only allowed from interrupt handling

Note: For library model, the SVC numbers have no restrictions.

Since the requirements for SVC number assignment are different,
this patch also split the SVC number header for IPC and Library models.

Change-Id: I0fb4dd110be6bab05e1c4b9a8fc55e1b8bfbc0eb
Signed-off-by: Kevin Peng <kevin.peng@arm.com>