psa-arch-tests:  cmake: api-tests: Allow CROSS_COMPILE to set C compiler
Cmake currently falls back to the default C compiler on the system,
which may not be an appropriate embedded toolchain in every case.
This commit checks for a `CROSS_COMPILE` field that can be
passed in to indicate the specific GCC cross-compiler to use. If no
matching value is found, it defaults to the system C compiler,
as was the case previously.

Submitted in github.com/ARM-software/psa-arch-tests/pull/268

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>