platform: stm: pass compiler path to STM32L552 postbuild script
The STM32 post-build script is passed the compiler full
name and path, and we remove the hard-coded call to
arm-none-eabi-gcc. The rationale for doing this is to
allow the script to work when the compiler application
is not in the (default) path, and/or to allow the script
to work with other GCC compilers (such as the Zephyr
SDK GCC compiler), not only with GMU ARM Embedded.

Change-Id: Ie45288002328bc03d0be0f4bbddd384f83741e36
Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>