SPM: Block calling certain SVC from FLIH Functions
FLIH Functions are not allowed to call certain SVC to avoid race
conditions. Check SVC_Handler_IPC() for the list of the SVCs banned.
When an FLIH Function is running, the Main Stack has contents in
addition to the stack seals.

This patch checks if Main Stack has extra contents and blocks calling
those SVCs from FLIH Functions.

Change-Id: Ifb10c3e31509624e0beaa99a84b3b84b2886a079
Signed-off-by: Kevin Peng <kevin.peng@arm.com>