platform: nordic: rearrange veneer placement on builds without BL2
For builds without BL2, or when BL2 will be booting a single
combined S and NS image, we force placing the veneers section
at the end of the image (position 3), so as not to waste space
as a result of the nRF veneer section alignment requirements.
For regular builds with BL2 we keep the veneer placement as is
(positions 1 or 2 depending on whether PSA_API_TEST_NS is
defined).

Change-Id: I292f9ec996445b9d7acedd1db24117d9345f3346
Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>