Platform: LPCxpresso55s69: Added TFM_PLATFORM_NXP_HAL_FILE_PATH variable
Added TFM_PLATFORM_NXP_HAL_FILE_PATH CMake variable for specifying
the location of NXP SDK hal. Defaultly the variable is set to DOWNLOAD.
In this case, the pulling drivers from git during build is enabled.
Otherwise, when specified the path to the checked out repository,
this funcionality is disabled and the files are linked statically.

Added NXP_SDK_GIT_TAG CMake variable for specifying the version
of NXP MCUXpresso SDK.

Change-Id: Ie3a6048d76f760591debbf9bee1c7c7c85a4b62d
Signed-off-by: MartinaHanusovaNXP <martina.hanusova@nxp.com>