SPM: Remove the IRQ priority in manifest
TF-M added a tfm_irq_priority attribute in the Partition manifest
for SPM to set IRQ priorities so that IRQ can work correctly - TF-M
requires external interrupt priorities must be higher than that of
PendSV.

However, a universal IRQ priority might not work on all platforms
because different platforms have different number of configurable
interrupt priorities. So the same value represents different
priorities on different platforms.
Besides, the attribute is not defined by FF-M.

This patch removes the IRQ priority in manifest and in the HAL API
as well so that platforms have the most flexibility to set priorities
for IRQs based on their own cases.

Note: the external interrupt priorities must be higher than PendSV.

Change-Id: Id9e544a9afffcc7d019177cf29e51f32d1600504
Signed-off-by: Kevin Peng <kevin.peng@arm.com>