SPM: Correct the PendSV priority value
Priority boost for Secure IRQ is enabled in TF-M - NS IRQ priorities
values are mapped to 0x80 - 0xFF (lower priority).
So the lowest priority value for SPE is 0x80.
It is not need to "-1" because exception preemption does not happen
between exceptions with same priority values.

The lowest priority exception is the PendSV in SPE.
This patch fixes its priority value by removing the "-1".

Change-Id: I1d340cda939f7785d684428bcbb54c7aea4445c4
Signed-off-by: Kevin Peng <kevin.peng@arm.com>