Docs: Fix the broken links in the documentation
Sphinx uses "autosectionlable" extension for generating the section
labels automatically. The max depth is set to 2 for level 2 headers
only. That section which is under level 2 will not have the reference
label. This caused the broken links in the documentation.

This patch creates lables for the section which is under level 2 and
referenced by other places in the documentation.

Signed-off-by: David Wang <david.wang@arm.com>
Change-Id: I636922c624108e850635e683719f5ac035e0880b