Tools: Silence Generating... messages
The build log is cluttered with messages that are usually
inconsequential. These messages start with 'Generating' and tend to
hide warnings and errors that appear.

This commit adds a quiet flag to the script in question and allows
enabling the quiet mode by with a flag. If you would not like to see
these messages enable PARSE_MANIFEST_QUIET. This goes a long way to
having a shorter, less noisy build log.

Change-Id: I085ac4f4909e86c3ccecf1bdf5482bafaecd3dcf
Signed-off-by: Jimmy Brisson <jimmy.brisson@linaro.org>