Allow Silencing most messages
Many users of TF-M don't need most of the build log, and would
like the build log to consist of errors and warnings only. This
commit introduces the TFM_BUILD_LOG_VERBOSE configuration variable
that disables most messages in the build log. In particular, this
configuration variable silences:
  * Installing messages
  * Generating messages
  * Partition configuration messages
  * Memory usage messages

All of these messages are on by default.

Signed-off-by: Jimmy Brisson <jimmy.brisson@linaro.org>
Change-Id: I7dd4e23d43e4a2a32c0189f99ea1818628ace32e