Tools: Auto-allocation PID starts from max fixed PID + 1
Usually manifests are listed in ascending order by PID.
This patch starts auto-allocation PID from the maximum fixed PIDs
so that the generated pid.h can have a ascending order with PIDs
as well.

This is sometimes helpful in debugging to know which type of Partition
a certain PID is, for example a test Partition or not.
Because Partition list is usually sorted by types as well.

Change-Id: Ie3b5a5af093140cddfb34337b34bacf96a5aae0b
Signed-off-by: Kevin Peng <kevin.peng@arm.com>